move message socket to Home stops it from removing event listener when component closes on mobile
