+++ src/components/Home/Home.js
@@ -1,7 +1,7 @@
 import React, { useState, useEffect } from 'react';
 import { connect, useDispatch } from 'react-redux';
 import Trade from '../Trade/Trade.js';
-import Updates from '../Messages/Messages.js';
+import Messages from '../Messages/Messages.js';
 import Menu from '../Menu/Menu'
 import TradeList from '../TradeList/TradeList'
 import Status from '../Status/Status'
@@ -13,16 +13,71 @@ import NotActive from '../NotActive/NotActive.js';
 import { SocketProvider } from '../../contexts/SocketProvider.js';
 import axios from 'axios';
 import MobileNav from '../MobileNav/MobileNav.js';
+import { useSocket } from "../../contexts/SocketProvider";
 import useWindowDimensions from '../../hooks/useWindowDimensions.js';
 
 function Home(props) {
   const dispatch = useDispatch();
   const { height, width } = useWindowDimensions();
 
+  const socket = useSocket();
+  const [messages, setMessages] = useState([]);
+  const [messagesCount, setMessagesCount] = useState(0);
+  const [errors, setErrors] = useState([]);
+  const [errorCount, setErrorCount] = useState(0);
+
   const [showSettings, setShowSettings] = useState(false);
   const [theme, setTheme] = useState('original');
   const [mobilePage, setMobilePage] = useState('tradeList');
   const [priceTicker, setPriceTicker] = useState(0);
+
+  // need use effect to prevent multiplying connections every time component renders
+  useEffect(() => {
+    // socket may not exist on page load because it hasn't connected yet
+    if (socket == null) return;
+    socket.on('message', message => {
+      // console.log('HERE IS THE WHOLE MESSAGE', message);
+      if (message.userID === props.store.accountReducer.userReducer.id) {
+        // console.log('HERE IS THE USER ID FROM THE MESSAGE', message.userID);
+        if (message.message) {
+          setMessagesCount(prevMessagesCount => {
+            return prevMessagesCount + 1;
+          });
+          setMessages(prevMessages => {
+            // keep max messages down to 3 by checking if more than 2 before adding new message
+            if (prevMessages.length > 999) {
+              prevMessages.pop();
+            }
+            let datedMessage = {
+              date: `${Date()}`,
+              message: `${message.message}`
+            }
+            return [datedMessage, ...prevMessages]
+          });
+        }
+        if (message.error) {
+          setErrorCount(prevErrorCount => {
+            return prevErrorCount + 1;
+          });
+          setErrors(prevErrors => {
+            // keep max messages down to 3 by checking if more than 2 before adding new message
+            if (prevErrors.length > 999) {
+              prevErrors.pop();
+            }
+            let datedError = {
+              date: `${Date()}`,
+              error: `${message.error}`
+            }
+            return [datedError, ...prevErrors]
+          });
+        }
+      }
+    });
+    // this will remove the listener when component rerenders
+    return () => socket.off('message')
+    // useEffect will depend on socket because the connection will 
+    // not be there right when the page loads
+  }, [socket, props.store.accountReducer.userReducer.id]);
 
   const clickSettings = () => {
     setShowSettings(!showSettings);
@@ -93,9 +148,9 @@ function Home(props) {
         }
 
         {width < 800 && mobilePage === 'messages'
-              ? <Updates theme={theme} />
-              : width > 800 && <Updates theme={theme} />}
-        
+          ? <Messages theme={theme} messages={messages} messagesCount={messagesCount} errors={errors} errorCount={errorCount} />
+          : width > 800 && <Messages theme={theme} messages={messages} messagesCount={messagesCount} errors={errors} errorCount={errorCount} />}
+
         <Status theme={theme} priceTicker={priceTicker} />
         <Settings showSettings={showSettings} clickSettings={clickSettings} theme={theme} priceTicker={priceTicker} />
         {width < 800 && <MobileNav theme={theme} setMobilePage={setMobilePage} />}
